pyproject.toml
[project]
name = "chemtools-manager"
dynamic = [ "version" ]
authors = [
    {name = "silox",email = "edruisl99@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "docker (>=7.1.0,<8.0.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "fastapi-utils (>=0.8.0,<0.9.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "typing-inspect (>=0.9.0,<0.10.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "gevent (>=24.11.1,<25.0.0)",
]

[tool.poetry]
version = "0.3.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
mypy = "^1.14.1"
pytest = "^8.3.4"
lxml-stubs = "^0.5.1"
types-docker = "^7.1.0.20241229"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
# Selected tools
# Pyflakes - F
# pycodestyle - E, W
# flake8-buildins - A
# isort - I
# pylint - PLC, PLE, PLR, PLW
select = ["F", "E", "W", "A", "I", "PLC", "PLE", "PLR", "PLW"]
ignore = ["PLR0913", "PLR2004"]
